Orientation component container is now a templatized VectorMap.
