+++ include/components/SigmaMotionState.h
@@ -4,30 +4,27 @@
 //#define BT_USE_DOUBLE_PRECISION
 #include "bullet/btBulletDynamicsCommon.h"
 #include "components/WorldPosition.h"
-#include "components/WorldOrientation.h"
+#include "VectorMap.hpp"
 #include <cmath>
 
 namespace Sigma {
+	struct orientation_type;
+
     class SigmaMotionState : public btMotionState {
     public:
-        SigmaMotionState(const id_t id, WorldPosition& wp, WorldOrientation& wo)
+        SigmaMotionState(const id_t id, WorldPosition& wp, VectorMap<id_t, orientation_type>& wo)
         : id(id), wp(&wp), wo(&wo) {};
 
         virtual ~SigmaMotionState() {};
 
-        virtual void getWorldTransform(btTransform &worldTrans) const {
-            worldTrans.setIdentity();
-            worldTrans.setOrigin(btVector3(wp->x(id), wp->y(id), wp->z(id)));
-            orientation_type o = wo->at(id);
-            worldTrans.getBasis().setEulerZYX(o.alpha, o.beta, o.gamma);
-        };
+        virtual void getWorldTransform(btTransform &worldTrans) const;
 
         virtual void setWorldTransform(const btTransform &worldTrans);
 
     private:
         const id_t id;
         WorldPosition* wp;
-        WorldOrientation* wo;
+        VectorMap<id_t, orientation_type>* wo;
     };
 }
