+++ tests/main.cpp
@@ -1,7 +1,6 @@
 #include "gtest/gtest.h"
-#include "tests/WorldOrientationTest.h"
-#include "tests/BitArrayTest.h"
 #include "tests/WorldPositionTest.h"
+#include "tests/BitArrayTest.h"
 #include "tests/VectorMapTest.h"
 #include "tests/PropertyTest.h"
